+++ MS5837.h
@@ -2,7 +2,7 @@
 //
 //    FILE: MS5837.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 //    DATE: 2023-11-12
 // PURPOSE: Arduino library for MS5837 temperature and pressure sensor.
 //     URL: https://github.com/RobTillaart/MS5837
@@ -12,13 +12,20 @@
 #include "Wire.h"
 
 
-#define MS5837_LIB_VERSION        (F("0.1.1"))
+#define MS5837_LIB_VERSION        (F("0.1.2"))
 
 
 //  TYPES
-#define MS5803_TYPE_01             1
-#define MS5837_TYPE_02             2
-#define MS5837_TYPE_30            30
+#define MS5837_TYPE_UNKNOWN       255
+#define MS5803_TYPE_01              1
+#define MS5837_TYPE_02              2
+#define MS5837_TYPE_30             30
+
+
+//  ERROR CODES
+#define MS5837_OK                 0x00
+//  TODO elaborate
+//  I2C can return 2, 3, 4
 
 
 class MS5837
@@ -93,7 +100,7 @@ protected:
   float     _temperature;
 
   float     C[8];
-  uint8_t   _type;
+  uint8_t   _type = MS5837_TYPE_UNKNOWN;
   int       _result;
 
   float     _density = 0.99802;  //  water at 20 °C